feat: Created home and system pages
